Use shared image processing context
For some reason not using the shared image processing context causes the shared GPUImageFramebufferCache of the GPUImageContext to build upp to a point where memory warnings appear and finally makes the app crash when exporting a video in 720p.

If this fix causes other errors one might investigate further why this is happening.